Hard-code DT for the LiteX+RocketChip SoC
The DTS file assumes LiteX+RocketChip is deployed to a nexys4ddr
or ecp5versa FPGA board (128MB RAM, UART base address 0x92001800).

NOTE 1: Ultimately, the DT blob should be embedded in the BIOS,
on the "gateware" side of the hardware/software divide; for now,
it is more convenient to develop and debug with the DT embedded
in the BBL.

NOTE 2: BBL should s/imac/imafdc/ if able & willing to emulate FP
(see https://github.com/riscv/riscv-pk/issues/166).

Signed-off-by: Gabriel Somlo <gsomlo@gmail.com>